#365 Add confirm for remove point method

+++ src/components/organisms/popup/map.vue
@@ -79,11 +79,13 @@ export default {
           icon: this.ICONS.delete,
           label: this.$t('general.remove'),
           method: () => {
-            communicates.showSuccess('Trwa usuwanie punktu...');
-            this.popup.hide();
-            this.$store.dispatch('event/removePoint', this.$store.getters['mapPopup/pointId'])
-              .then(() => communicates.showSuccessTemporary(translator.t('communicate.map.pointRemoved')))
-              .catch(em => em.showMessage());
+            if (confirm(translator.t('communicate.map.confirmPointRemove'))) {
+              communicates.showSuccess(translator.t('communicate.map.pointRemovingInProgress'));
+              this.popup.hide();
+              this.$store.dispatch('event/removePoint', this.$store.getters['mapPopup/pointId'])
+                .then(() => communicates.showSuccessTemporary(translator.t('communicate.map.pointRemoved')))
+                .catch(em => em.showMessage());
+            }
           },
         },
       ];
